Moved specific utiities to their own component
