+++ src/App.vue
@@ -18,6 +18,7 @@ import Editor from "./components/Editor.vue";
 import Roadmap from "./components/Roadmap.vue";
 import StyleEditor from "./components/StyleEditor.vue";
 import Utilities from "./utilities/general";
+import IGraphic from "./models/IGraphic";
 
 @Component({
     components: {
@@ -49,8 +50,38 @@ export default class App extends Vue {
             }
         });
 
-        document.addEventListener("copy", Utilities.copyHandler(this));
-        document.addEventListener("paste", Utilities.pasteHandler(this));
+        // Overrides the default behavior of copy to copy the graphic model of the focused graphic
+        document.addEventListener("copy", (event: Event): void => {
+            // Cast event as clipboard event and prevent from copying any user selection
+            const clipboardEvent: ClipboardEvent = event as ClipboardEvent;
+            clipboardEvent.preventDefault();
+
+            const focusedGraphic: IGraphic | undefined = this.$store.getters.focusedGraphic;
+            if (focusedGraphic === undefined) {
+                return;
+            }
+
+            // Set the clipboard data to the graphic model associated with the current focused graphic
+            const graphic: IGraphic = this.$store.getters.activeSlide.graphics.find((graphic: IGraphic) => graphic.id === focusedGraphic.id)!;
+            clipboardEvent.clipboardData.setData("text/json", JSON.stringify(graphic));
+        });
+
+        document.addEventListener("paste", (event: Event): void => {
+            // Cast event as clipboard event
+            const clipboardEvent: ClipboardEvent = event as ClipboardEvent;
+            clipboardEvent.preventDefault();
+
+            const data: any = JSON.parse(clipboardEvent.clipboardData.getData("text/json"));
+            if (data === undefined) {
+                return;
+            }
+
+            const graphic: IGraphic = Utilities.parseGraphic(data);
+            graphic.id = Utilities.generateId();
+            this.$store.commit("addGraphic", { slideId: this.$store.getters.activeSlide.id, graphic: graphic });
+            this.$store.commit("focusGraphic", graphic);
+            this.$store.commit("styleEditorObject", graphic);
+        });
     }
 
     get workspaceStyle(): any {
